chore: move handler util macros
