+++ server/Cargo.toml
@@ -10,7 +10,6 @@ vex-orderbook = { workspace = true }
 processors = { workspace = true }
 vex-config = { workspace = true }
 vex-networking = { workspace = true }
-utils = { workspace = true }
 parking_lot = {workspace = true}
 tokio = { version = "1", features = ["full"] }
 tracing = "0.1"
