+++ server/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod engine;
+pub mod events;
 pub mod utils;
 
 use std::sync::Arc;
