+++ server/src/utils.rs
@@ -8,60 +8,12 @@ macro_rules! create_risk_handler {
         let risk_engines_clone = $risk_engines.clone();
         move |cmd: &OrderCommand, _sequence: i64, _end_of_batch: bool| {
             let mut engine = risk_engines_clone[$shard_id].lock();
+            let mut cmd_clone = cmd.clone();
 
-            if let Err(e) = engine.pre_process_command(cmd) {
+            if let Err(e) = engine.pre_process_command(&mut cmd_clone) {
                 warn!("[RiskEngine_{}] Risk check failed: {:?}", $shard_id, e);
                 return;
             }
-        }
-    }};
-}
-
-/// Macro to generate risk engine R2 handlers (sharded by user_id)
-/// Each handler only processes events for users that belong to its shard
-#[macro_export]
-macro_rules! create_risk_r2_handler {
-    ($shard_id:expr, $risk_engines:expr) => {{
-        let risk_engines_clone = $risk_engines.clone();
-        move |event: &MatcherTradeEvent, _sequence: i64, _end_of_batch: bool| {
-            let num_shards = risk_engines_clone.len() as u64;
-            let shard_mask = num_shards - 1;
-
-            // Route to risk engine shard for active order user
-            let active_order_user_id = event.active_order_user_id;
-            let active_order_shard = (active_order_user_id & shard_mask) as usize;
-
-            // Only process if this event belongs to our shard
-            if active_order_shard == $shard_id {
-                if let Some(risk_engine_mutex) = risk_engines_clone.get($shard_id) {
-                    let mut risk_engine = risk_engine_mutex.lock();
-                    risk_engine.handle_event(event);
-                }
-            }
-
-            // Route to risk engine shard for maker user (if different)
-            let maker_user_id = event.maker_user_id;
-            if maker_user_id != active_order_user_id {
-                let maker_shard = (maker_user_id & shard_mask) as usize;
-
-                // Only process if this event belongs to our shard
-                if maker_shard == $shard_id {
-                    if let Some(risk_engine_mutex) = risk_engines_clone.get($shard_id) {
-                        let mut risk_engine = risk_engine_mutex.lock();
-                        risk_engine.handle_event(event);
-                    }
-                }
-            }
-        }
-    }};
-}
-
-#[macro_export]
-macro_rules! create_event_handler {
-    ($events_handler:expr) => {{
-        let events_handler = $events_handler.clone();
-        move |event: &MatcherTradeEvent, _sequence: i64, _end_of_batch: bool| {
-            events_handler.handle_event(event.clone());
         }
     }};
 }
@@ -70,25 +22,57 @@ macro_rules! create_event_handler {
 /// This eliminates code duplication while maintaining separate handlers for each shard
 #[macro_export]
 macro_rules! create_matching_handler {
-    ($shard_id:expr, $routers:expr, $matcher_event_producer:expr) => {{
-        use disruptor::Producer;
+    ($shard_id:expr, $routers:expr, $events_handler:expr, $journaling:expr, $risk_engines:expr) => {{
         let router = $routers[$shard_id].clone();
-        let mut matcher_event_producer = $matcher_event_producer.clone();
+        let events_handler = $events_handler.clone();
+        let journaling = $journaling.clone();
+        let risk_engines = $risk_engines.clone();
 
-        move |cmd: &OrderCommand, _sequence: i64, _end_of_batch: bool| {
-            // Only lock during order processing
-            let processed_order_cmd = {
-                let mut router_guard = router.lock();
-                // remove this lock eventually by some minor optimsations in orderbook and matching engine router
-                let mut order_cmd = cmd.clone();
-                let processed_order_cmd = router_guard.process_order(&mut order_cmd);
-                processed_order_cmd
-            };  // Lock is released here - router is free for next order
+        move |cmd: &OrderCommand, _sequence: u64, _end_of_batch: bool| {
+            // Lock the specific matching engine router shard
+            let mut router_guard = router.lock();
+            let mut cmd_clone = cmd.clone();
 
-            // Publish raw events directly
-            let _ = matcher_event_producer.publish(|published_event| {
-                *published_event = processed_order_cmd;
-            });
+            // Each router filters internally based on symbol_id ownership
+            // This ensures only the correct shard processes each symbol_id's orders
+            router_guard.process_order(&mut cmd_clone);
+
+            // Process all events generated by this matching engine
+            // Events represent trades, cancellations, etc.
+            let mut current_event = cmd_clone.matcher_event.take();
+            while let Some(event_box) = current_event {
+                let mut event = *event_box;
+                current_event = event.next_event.take();
+
+                // Journal the event for audit trail and recovery
+                journaling.journal_event(&event);
+
+                // Route events to correct risk engine shards for settlement (R2)
+                // Events can affect multiple users (maker and taker), so route to both
+                let num_shards = risk_engines.len() as u64;
+                let shard_mask = num_shards - 1; // Power of 2 for efficient bitwise ops
+
+                // Route to risk engine shard for active order user
+                let active_order_user_id = event.active_order_user_id;
+                let active_order_shard = (active_order_user_id & shard_mask) as usize;
+                if let Some(risk_engine_mutex) = risk_engines.get(active_order_shard) {
+                    let mut risk_engine = risk_engine_mutex.lock();
+                    risk_engine.handle_event(&event); // Risk release (R2)
+                }
+
+                // Route to risk engine shard for maker user (if different)
+                let maker_user_id = event.maker_user_id;
+                if maker_user_id != active_order_user_id {
+                    let maker_shard = (maker_user_id & shard_mask) as usize;
+                    if let Some(risk_engine_mutex) = risk_engines.get(maker_shard) {
+                        let mut risk_engine = risk_engine_mutex.lock();
+                        risk_engine.handle_event(&event); // Risk release (R2)
+                    }
+                }
+
+                // Send event to external handlers (market data, notifications, etc.)
+                events_handler.handle_event(event.clone());
+            }
         }
     }};
 }
