--- utils/src/handler_util.rs
@@ -1,78 +1,0 @@
-// Macros for creating handlers for risk and matching engines
-
-/// Macro to generate risk engine handlers
-/// This eliminates code duplication while maintaining separate handlers for each shard
-#[macro_export]
-macro_rules! create_risk_handler {
-    ($shard_id:expr, $risk_engines:expr) => {{
-        let risk_engines_clone = $risk_engines.clone();
-        move |cmd: &OrderCommand, _sequence: i64, _end_of_batch: bool| {
-            let mut engine = risk_engines_clone[$shard_id].lock();
-            let mut cmd_clone = cmd.clone();
-
-            if let Err(e) = engine.pre_process_command(&mut cmd_clone) {
-                warn!("[RiskEngine_{}] Risk check failed: {:?}", $shard_id, e);
-                return;
-            }
-        }
-    }};
-}
-
-/// Macro to generate matching engine handlers
-/// This eliminates code duplication while maintaining separate handlers for each shard
-#[macro_export]
-macro_rules! create_matching_handler {
-    ($shard_id:expr, $routers:expr, $events_handler:expr, $journaling:expr, $risk_engines:expr) => {{
-        let router = $routers[$shard_id].clone();
-        let events_handler = $events_handler.clone();
-        let journaling = $journaling.clone();
-        let risk_engines = $risk_engines.clone();
-
-        move |cmd: &OrderCommand, _sequence: u64, _end_of_batch: bool| {
-            // Lock the specific matching engine router shard
-            let mut router_guard = router.lock();
-            let mut cmd_clone = cmd.clone();
-
-            // Each router filters internally based on symbol_id ownership
-            // This ensures only the correct shard processes each symbol_id's orders
-            router_guard.process_order(&mut cmd_clone);
-
-            // Process all events generated by this matching engine
-            // Events represent trades, cancellations, etc.
-            let mut current_event = cmd_clone.matcher_event.take();
-            while let Some(event_box) = current_event {
-                let mut event = *event_box;
-                current_event = event.next_event.take();
-
-                // Journal the event for audit trail and recovery
-                journaling.journal_event(&event);
-
-                // Route events to correct risk engine shards for settlement (R2)
-                // Events can affect multiple users (maker and taker), so route to both
-                let num_shards = risk_engines.len() as u64;
-                let shard_mask = num_shards - 1; // Power of 2 for efficient bitwise ops
-
-                // Route to risk engine shard for active order user
-                let active_order_user_id = event.active_order_user_id;
-                let active_order_shard = (active_order_user_id & shard_mask) as usize;
-                if let Some(risk_engine_mutex) = risk_engines.get(active_order_shard) {
-                    let mut risk_engine = risk_engine_mutex.lock();
-                    risk_engine.handle_event(&event); // Risk release (R2)
-                }
-
-                // Route to risk engine shard for maker user (if different)
-                let maker_user_id = event.maker_user_id;
-                if maker_user_id != active_order_user_id {
-                    let maker_shard = (maker_user_id & shard_mask) as usize;
-                    if let Some(risk_engine_mutex) = risk_engines.get(maker_shard) {
-                        let mut risk_engine = risk_engine_mutex.lock();
-                        risk_engine.handle_event(&event); // Risk release (R2)
-                    }
-                }
-
-                // Send event to external handlers (market data, notifications, etc.)
-                events_handler.handle_event(event.clone());
-            }
-        }
-    }};
-}
